refactor(basic): split semantic analyzer into passes
